\name{countdata}
\alias{countdata}
\docType{data}
\title{
Example data 
}
\description{
Randomly generated example data representing a count matrix
}
\usage{data(exampleData)}
\details{
The data consists of a variable \code{countdata} with a matrix with 1113 genes and 96 samples. The count values are randomly generated numbers.
}
\examples{
data(exampleData)

print(countdata)

}