\name{labels}
\alias{labels}
\docType{data}
\title{
Example data to label samples in countdata (rows) in two classes 0 and 1
}
\description{
Randomly generated example data to give samples in countdata a class label 0 or 1
}
\usage{data(exampleData)}
\details{
The data consists of a variable \code{labels} with a vector of 96 values. The values are a random sequence of 0's and 1's.
}
\examples{
data(exampleData)

print(labels)

}